pyproject.toml
[project]
name = "decision-maker-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "fastapi>=0.116.1",
    "firebase-admin>=7.1.0",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]

[project.optional-dependencies]
dev = [
    "black>=25.1.0",
    "mypy>=1.17.1",
    "pytest>=8.4.1",
    "types-firebase-admin>=0.1.1",
]
